Fix GnGeo driver ROM entry alignment

+++ src/gngeo/rom_types.h
@@ -12,7 +12,7 @@ typedef uint32_t Uint32;
 #define REGION_AUDIO_CPU_ENCRYPTED   2
 #define REGION_AUDIO_DATA_1          3
 #define REGION_AUDIO_DATA_2          4
-#define REGION_FIXED_LAYER_BIOS     5
+#define REGION_FIXED_LAYER_BIOS      5
 #define REGION_FIXED_LAYER_CARTRIDGE 6
 #define REGION_MAIN_CPU_BIOS         7
 #define REGION_MAIN_CPU_CARTRIDGE    8
@@ -40,51 +40,51 @@ typedef enum COUNTRY {
 } COUNTRY;
 
 typedef struct ROM_DEF {
-	char name[32];
-	char parent[32];
-	char longname[128];
-	Uint32 year;
-	Uint32 romsize[10];
-	Uint32 nb_romfile;
+    char name[32];
+    char parent[32];
+    char longname[128];
+    Uint32 year;
+    Uint32 romsize[10];
+    Uint32 nb_romfile;
 
-	struct romfile {
-		char filename[32];
-		Uint8 region;
-		Uint32 src;
-		Uint32 dest;
-		Uint32 size;
-		Uint32 crc;
-	} rom[32];
+    struct romfile {
+        char filename[32];
+        Uint32 region;
+        Uint32 src;
+        Uint32 dest;
+        Uint32 size;
+        Uint32 crc;
+    } rom[32];
 } ROM_DEF;
 
 typedef struct GAME_INFO {
-	char *name;
-	char *longname;
-	int year;
-	Uint32 flags;
+    char *name;
+    char *longname;
+    int year;
+    Uint32 flags;
 } GAME_INFO;
 
 typedef struct ROM_REGION {
-	Uint8 *p;
-	Uint32 size;
+    Uint8 *p;
+    Uint32 size;
 } ROM_REGION;
 
 typedef struct GAME_ROMS {
-	GAME_INFO info;
-	ROM_REGION cpu_m68k;
-	ROM_REGION cpu_z80;
-	ROM_REGION tiles;
-	ROM_REGION game_sfix;
-	ROM_REGION bios_sfix;
-	ROM_REGION bios_audio;
-	ROM_REGION zoom_table;
-	ROM_REGION bios_m68k;
-	ROM_REGION adpcma;
-	ROM_REGION adpcmb;
-	ROM_REGION spr_usage;
-	ROM_REGION gfix_usage;
-	ROM_REGION bfix_usage;
-	ROM_REGION cpu_z80c;
+    GAME_INFO info;
+    ROM_REGION cpu_m68k;
+    ROM_REGION cpu_z80;
+    ROM_REGION tiles;
+    ROM_REGION game_sfix;
+    ROM_REGION bios_sfix;
+    ROM_REGION bios_audio;
+    ROM_REGION zoom_table;
+    ROM_REGION bios_m68k;
+    ROM_REGION adpcma;
+    ROM_REGION adpcmb;
+    ROM_REGION spr_usage;
+    ROM_REGION gfix_usage;
+    ROM_REGION bfix_usage;
+    ROM_REGION cpu_z80c;
 } GAME_ROMS;
 
 #endif
